+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Profiler Change Log
 
-
 ## 1.0.2 under development
 
-- no changes in this release.
+- Bug #42: Forbid passing context with "beginMemory" or "beginTime" keys to `Profiler::end()` (vjik)
+- Bug #42: Check that category of a message is a string (vjik)
 
 ## 1.0.1 November 12, 2021